+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a 404 page [#51](https://github.com/azavea/fb-gender-survey-dashboard/pull/51)
 - Add search for countries and questions [#50](https://github.com/azavea/fb-gender-survey-dashboard/pull/50)
+- Allow Question Category Select-All/None [#68](https://github.com/azavea/fb-gender-survey-dashboard/pull/68)
 
 ### Changed